improved templating for create script

+++ create-bolt-uxp/src/build.ts
@@ -5,16 +5,12 @@ import * as color from "picocolors";
 
 import * as prettier from "prettier";
 
-import { appOptions, frameworkOptions } from ".";
+import { appOptions, frameworkOptions } from "./data";
 import { execAsync, posix } from "./utils";
 import { spinner, note } from "@clack/prompts";
 
 import { capitalize } from "radash";
 
-export type Opt = {
-  value: string;
-  label: string;
-};
 type Args = {
   folder: string;
   displayName: string;
@@ -28,12 +24,15 @@ type Args = {
 
 const multiBlankLineRegex = /(\r?\n\s*){1,}/g;
 
-const getJSRegex = (variable: string) =>
+const getJSRangeRegex = (variable: string) =>
   new RegExp(
-    `\/\/ BOLT-UXP_${variable}_START[\\s\\S]*?\/\/ BOLT-UXP_${variable}_END.*`,
-    "g"
+    `\/\/ BOLT-UXP_${variable}_START[\\s\\S]*?\/\/ BOLT-UXP_${variable}_END.*(\n|\r\n)?`,
+    "gm"
   );
-const allCommentsRegex = /\/\/ BOLT-UXP_.*_(START|END)/g;
+const getJSOnlyRegex = (variable: string) =>
+  new RegExp(`^.*\/\/ BOLT-UXP_${variable}_ONLY.*(\n|\r\n)?`, "gm");
+
+const allCommentsRegex = /\/\/ BOLT-UXP_.*_(START|END|ONLY).*(\n|\r\n)?/gm;
 
 const getHTMLRegex = (variable: string) =>
   new RegExp(
@@ -41,13 +40,61 @@ const getHTMLRegex = (variable: string) =>
     "g"
   );
 
-const allHTMLCommentsRegex = /<!-- BOLT-UXP_.*_(START|END) -->/g;
+const allHTMLCommentsRegex =
+  /<!-- BOLT-UXP_.*_(START|END|ONLY) -->.*(\n|\r\n)?/gm;
 const htmlDisabledScriptTagRegexStart = /<!-- <script/g;
 const htmlDisabledScriptTagRegexEnd = /<\/script> -->.*/g;
 
-export const buildBoltUXP = async (args: Args) => {
-  // console.log(args);
+const formatFile = async (
+  txt: string,
+  ext: string,
+  {
+    enableHybrid,
+    removeApps,
+    removeFrameworks,
+  }: {
+    enableHybrid: boolean;
+    removeApps: string[];
+    removeFrameworks: string[];
+  }
+) => {
+  [...removeApps, ...removeFrameworks].map((app) => {
+    const upper = app.toUpperCase();
 
+    const rangeRegex = getJSRangeRegex(upper);
+    const onlyRegex = getJSOnlyRegex(upper);
+    const rangeRegexHTML = getHTMLRegex(upper);
+    const onlyRegexHTML = getHTMLRegex(upper);
+
+    txt = txt
+      .replace(rangeRegex, "")
+      .replace(onlyRegex, "")
+      .replace(rangeRegexHTML, "")
+      .replace(onlyRegexHTML, "");
+  });
+  if (!enableHybrid) {
+    const rangeRegex = getJSRangeRegex("HYBRID");
+    const onlyRegex = getJSOnlyRegex("HYBRID");
+    txt = txt.replace(rangeRegex, "").replace(onlyRegex, "");
+  }
+  // cleanup
+  txt = txt.replace(allCommentsRegex, "");
+  txt = txt.replace(allHTMLCommentsRegex, "");
+  if (ext === ".html") {
+    txt = txt.replace("<!-- Uncomment to debug the desired template -->", "");
+    // txt = txt.replace(multiBlankLineRegex, "\n");
+    // re-enable all disabled <script /> tags
+    txt = txt
+      .replace(htmlDisabledScriptTagRegexStart, "<script")
+      .replace(htmlDisabledScriptTagRegexEnd, "</script>");
+  }
+  //   txt = await prettier.format(txt, {
+  //     parser: "typescript",
+  //   });
+  return txt;
+};
+
+export const buildBoltUXP = async (args: Args) => {
   const fullPath = path.join(process.cwd(), args.folder);
   note(`Creating Bolt UXP in ${color.green(color.bold(fullPath))}`, "Info");
 
@@ -57,53 +104,23 @@ export const buildBoltUXP = async (args: Args) => {
   const globalStem = posix(path.join(__dirname, `../../bolt-uxp`));
   const stem = fs.existsSync(globalStem) ? globalStem : localStem;
 
-  // note(
-  //   `Local Stem ${color.green(color.bold(localStem))} exists ${fs.existsSync(
-  //     localStem
-  //   )}`,
-  //   "Info"
-  // );
-  // note(
-  //   `Global Stem ${color.green(color.bold(globalStem))} exists ${fs.existsSync(
-  //     globalStem
-  //   )}`,
-  //   "Info"
-  // );
-  // note(`Copying files from ${color.green(color.bold(stem))}`, "Info");
-
-  if (fs.existsSync(fullPath)) {
-    fs.rmSync(fullPath, { recursive: true });
-  }
-
+  if (fs.existsSync(fullPath)) fs.rmSync(fullPath, { recursive: true });
   fs.mkdirSync(fullPath, { recursive: true });
-
-  const boltUxpFolder = path.join(__dirname, "..", "node_modules", "bolt-uxp");
-  // console.log(boltUxpFolder);
 
   let includes: string[] = ["*", "src/**/*", "public/**/*", "public-zip/**/*"];
   let excludes: string[] = [];
 
-  const reactFiles = [
-    "src/index-react.tsx",
-    "src/main.tsx",
-    "package.react.json",
+  excludes = [
+    ...excludes,
+    ...frameworkOptions
+      .filter((f) => f.value !== args.framework)
+      .map((f) => f.files)
+      .flat(),
+    ...appOptions
+      .filter((f) => !args.apps.includes(f.value))
+      .map((f) => f.files)
+      .flat(),
   ];
-
-  const svelteFiles = [
-    "src/index-svelte.ts",
-    "src/main.svelte",
-    "package.svelte.json",
-  ];
-
-  const vueFiles = ["src/index-vue.ts", "src/main.vue", "package.vue.json"];
-
-  if (args.framework === "react") {
-    excludes = [...excludes, ...svelteFiles, ...vueFiles];
-  } else if (args.framework === "svelte") {
-    excludes = [...excludes, ...reactFiles, ...vueFiles];
-  } else if (args.framework === "vue") {
-    excludes = [...excludes, ...reactFiles, ...svelteFiles];
-  }
 
   if (args.enableHybrid) {
     includes = [...includes, "public-hybrid/**/*"];
@@ -118,20 +135,17 @@ export const buildBoltUXP = async (args: Args) => {
     ],
     {
       onlyFiles: true,
-      followSymbolicLinks: true, // Set to false to not follow symlinks
+      followSymbolicLinks: true,
     }
   );
-  // note(
-  //   `Copying ${color.green(
-  //     color.bold(files.length.toString())
-  //   )} files to ${color.green(color.bold(fullPath))}`,
-  //   "Info"
-  // );
-  // note(files.map((file) => file.replace(stem, "")).join("\n"), "Info");
-  // console.log(files);
 
-  files.map((file) => {
-    // const fileName = path.basename(file);
+  const allApps = appOptions.map((app) => app.value);
+  const removeApps = allApps.filter((app) => !args.apps.includes(app));
+  const removeFrameworks = frameworkOptions
+    .filter((f) => f.value !== args.framework)
+    .map((f) => f.value);
+
+  for (const file of files) {
     const fileName = file.replace(stem, "");
     const dest = path.join(fullPath, fileName);
     const parent = path.dirname(dest);
@@ -142,8 +156,22 @@ export const buildBoltUXP = async (args: Args) => {
       });
     } else {
       fs.copyFileSync(file, dest);
+      const txt = fs.readFileSync(dest, "utf8");
+      const ext = path.extname(dest);
+
+      const newTxt = await formatFile(txt, ext, {
+        enableHybrid: args.enableHybrid,
+        removeApps,
+        removeFrameworks,
+      });
+
+      // wite file if changed
+      if (newTxt !== txt) {
+        console.log(`UPDATING CHANGED: ${color.green(color.bold(fileName))}`);
+        fs.writeFileSync(dest, newTxt, "utf8");
+      }
     }
-  });
+  }
 
   //* Update Config
 
@@ -166,104 +194,14 @@ export const buildBoltUXP = async (args: Args) => {
   //* update uxp.config.ts
   const uxpConfig = path.join(fullPath, "uxp.config.ts");
   let uxpConfigData = fs.readFileSync(uxpConfig, "utf8");
-
-  // update name
-  const nameRegex = /name: \".*\",/;
-  uxpConfigData = uxpConfigData.replace(
-    nameRegex,
-    `name: "${args.displayName}",`
-  );
-
-  // update id
-  const idRegex = /id: \".*\",/;
-  uxpConfigData = uxpConfigData.replace(idRegex, `id: "${args.id}",`);
-
-  // update apps
-  const allApps = appOptions.map((app) => app.value);
-  const removeApps = allApps.filter((app) => !args.apps.includes(app));
-  removeApps.map((app) => {
-    const upper = app.toUpperCase();
-    const regex = getJSRegex(upper);
-    // console.log(regex);
-    uxpConfigData = uxpConfigData.replace(regex, "");
-  });
-
-  // update hybrid
-  if (!args.enableHybrid) {
-    const regex = getJSRegex("HYBRID");
-    uxpConfigData = uxpConfigData.replace(regex, "");
-  }
-
-  // cleanup
-  uxpConfigData = uxpConfigData.replace(allCommentsRegex, "");
-
-  const uxpConfigDataFormatted = !args.pretty
-    ? uxpConfigData
-    : await prettier.format(uxpConfigData, {
-        parser: "typescript",
-      });
-  fs.writeFileSync(uxpConfig, uxpConfigDataFormatted, "utf8");
-
-  //* update index.html
-
-  const indexHtml = path.join(fullPath, "index.html");
-  let indexHtmlData = fs.readFileSync(indexHtml, "utf8");
-
-  // update framework
-  const excludedFrameworks = frameworkOptions
-    .filter((f) => f.value !== args.framework)
-    .map((f) => f.value);
-
-  excludedFrameworks.map((framework) => {
-    const regex = getHTMLRegex(framework.toUpperCase());
-    // console.log(regex);
-    indexHtmlData = indexHtmlData.replace(regex, "");
-  });
-
-  // re-enable all disabled tags
-  indexHtmlData = indexHtmlData.replace(
-    htmlDisabledScriptTagRegexStart,
-    "<script"
-  );
-  indexHtmlData = indexHtmlData.replace(
-    htmlDisabledScriptTagRegexEnd,
-    "</script>"
-  );
-
-  // cleanup
-  indexHtmlData = indexHtmlData.replace(allHTMLCommentsRegex, "");
-  indexHtmlData = indexHtmlData.replace(
-    "<!-- Uncomment to debug the desired template -->",
-    ""
-  );
-  indexHtmlData = indexHtmlData.replace(multiBlankLineRegex, "\n");
-
-  const indexHtmlFormatted = !args.pretty
-    ? indexHtmlData
-    : await prettier.format(indexHtmlData, {
-        parser: "html",
-      });
-  fs.writeFileSync(indexHtml, indexHtmlFormatted, "utf8");
-
-  const viteConfig = path.join(fullPath, "vite.config.ts");
-  let viteConfigData = fs.readFileSync(viteConfig, "utf8");
-
-  excludedFrameworks.map((framework) => {
-    const regex = getJSRegex(framework.toUpperCase());
-    viteConfigData = viteConfigData.replace(regex, "");
-  });
-
-  // cleanup
-  viteConfigData = viteConfigData.replace(allCommentsRegex, "");
-  const viteConfigFormatted = !args.pretty
-    ? viteConfigData
-    : await prettier.format(viteConfigData, {
-        parser: "typescript",
-      });
-  fs.writeFileSync(viteConfig, viteConfigFormatted, "utf8");
+  uxpConfigData = uxpConfigData
+    // update name
+    .replace(/name: \".*\",/, `name: "${args.displayName}",`)
+    // update id
+    .replace(/id: \".*\",/, `id: "${args.id}",`);
+  fs.writeFileSync(uxpConfig, uxpConfigData, "utf8");
 
   // * Dependencies
-
   if (args.installDeps) {
     const s = spinner();
     s.start("Installing dependencies...");
